requirements.txt
streamlit
transformers
torch
Pillow
gTTS
textstat
scikit-learn
sentencepiece
numpy

